Deleted some thread sleeps :)
